# internal/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "开源、自托管轻量级发布平台 Ech0 的 API 文档",
        "title": "Ech0 API 文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:6277",
    "basePath": "/api",
    "paths": {
...
        "/audios/upload": {
            "post": {
                "description": "用户上传音频文件，成功后返回音频的访问 URL",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通用功能"
                ],
                "summary": "上传音频",
                "parameters": [
                    {
                        "type": "file",
                        "description": "音频文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/images/delete": {
            "delete": {
                "description": "用户删除已上传的图片，需传入图片 URL 和来源信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通用功能"
                ],
                "summary": "删除图片",
                "parameters": [
                    {
                        "description": "图片删除请求体",
                        "name": "imageDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ImageDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
        "/images/upload": {
            "post": {
                "description": "用户上传图片，成功后返回图片的访问 URL",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通用功能"
                ],
                "summary": "上传图片",
                "parameters": [
                    {
                        "type": "file",
                        "description": "图片文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/s3/presign": {
            "put": {
                "description": "获取用于上传文件到 S3 的预签名 URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通用功能"
                ],
                "summary": "获取 S3 预签名 URL",
                "parameters": [
                    {
                        "description": "S3 预签名请求体",
                        "name": "s3Dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetPresignURLDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取预签名 URL 失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/website/title": {
            "get": {
                "description": "获取网站标题",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "通用功能"
                ],
                "summary": "获取网站标题",
                "parameters": [
                    {
                        "type": "string",
                        "description": "网站URL",
                        "name": "website_url",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取网站标题失败",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code 状态码，非0时表示自定义HTTP业务状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "Data 响应数据，具体内容因接口而异"
                },
                "msg": {
                    "description": "Msg 返回信息，通常是状态描述",
                    "type": "string"
                }
            }
        },
...
        "model.GetPresignURLDto": {
            "type": "object",
            "required": [
                "file_name"
            ],
            "properties": {
                "content_type": {
                    "description": "文件的 MIME 类型",
                    "type": "string"
                },
                "file_name": {
                    "description": "原始文件名",
                    "type": "string"
                }
            }
        },
...
        "model.ImageDto": {
            "type": "object",
            "required": [
                "source",
                "url"
            ],
            "properties": {
                "height": {
                    "description": "图片高度",
                    "type": "integer"
                },
                "object_key": {
                    "description": "对象存储的 Key, 用于删除 S3/R2 上的图片",
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "url": {
                    "description": "图片的 URL 地址",
                    "type": "string"
                },
                "width": {
                    "description": "图片宽度",
                    "type": "integer"
                }
            }
        },
...
    }
}